+++ src/index.ts
@@ -2,6 +2,8 @@ export * from './drivers/ConfigurityDriver'
 export * from './drivers/EnvironmentDriver'
 export * from './drivers/LaunchDarklyServerDriver'
 export * from './drivers/SimpleKeyValueDriver'
+export * from './drivers/LaunchDarklyClientDriver'
+
 export * from './base-drivers/SyncFeatureManagerDriver'
 export * from './base-drivers/AsyncFeatureManagerDriver'