+++ src/types/common.types.ts
@@ -1,6 +1,9 @@
 /**
  * Common optional parameters for retrieving a flag.
  */
+/**
+ * Common optional parameters for retrieving a flag.
+ */
 export type CommonValueParams<Flags, K extends keyof Flags> = {
   /**
    * The default value to use if the flag is not found.